Add complement logic parameter and conciliation

+++ ParameterControl.Policy.DataAccess/ModifyPolicy.cs
@@ -34,7 +34,7 @@ namespace ParameterControl.Policy.DataAccess
                             command.Parameters.Add(new OracleParameter("PARAM_DESCRIPTION", entity.Description));
                             command.Parameters.Add(new OracleParameter("PARAM_MODIFIELDBY", entity.ModifieldBy));
                             command.Parameters.Add(new OracleParameter("PARAM_OBJETIVO", entity.Objetive));
-                            command.Parameters.Add(new OracleParameter("PARAM_ESTADO_ACTIVACION", Convert.ToInt16(entity.State)));
+                            command.Parameters.Add(new OracleParameter("PARAM_ESTADO_ACTIVACION", Convert.ToInt32(entity.State)));
                             OracleDataReader reader = command.ExecuteReader();
                             return response = 1;
                         }
