+++ ParameterControl.Policy.DataAccess/SetPolicy.cs
@@ -33,7 +33,7 @@ namespace ParameterControl.Policy.DataAccess
                             command.Parameters.Add(new OracleParameter("PARAM_DESCRIPTION", entity.Description));
                             command.Parameters.Add(new OracleParameter("PARAM_MODIFIELDBY", entity.ModifieldBy));
                             command.Parameters.Add(new OracleParameter("PARAM_OBJETIVO", entity.Objetive));
-                            command.Parameters.Add(new OracleParameter("PARAM_ESTADO_ACTIVACION", Convert.ToUInt16(entity.State)));
+                            command.Parameters.Add(new OracleParameter("PARAM_ESTADO_ACTIVACION", Convert.ToInt32(entity.State)));
                             OracleDataReader reader = command.ExecuteReader();
                             return response = 1;
                         }
